Time: 0 ms (100.00%) | Memory: 8.1 MB (53.09%) - LeetSync

+++ 172-factorial-trailing-zeroes/factorial-trailing-zeroes.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int trailingZeroes(int n) {
-        long long noF=0;
+        int noF=0;
         for(int i=1;i<=n;i++)
         {
             int n=i;
